Add details of removeElement() method
I found `removeElement()` while looking for a practical solution to the issue with the EE Model `Search` method - which does bi-directional wildcard searches on the term given rather than literal searches - which in some instances will result in a collection being returned that has too many entries (i.e. you get all those which **contain** the search term in a field rather than **match** the search term).  

Running a foreach() loop on the output testing the 'seached' field against the original search term allows the incorrect elements to be indentified, **removeElement()** allows them to be removed from the collection.

**removeElement()** is one of several public methods in the Collection Class that are not documented.